% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building_toolkit.R
\name{build_alpha_p_inits}
\alias{build_alpha_p_inits}
\title{Build list of initial values for detection intercept(s)}
\usage{
build_alpha_p_inits(n_datasets, init_value = 0.5)
}
\arguments{
\item{n_datasets}{Number of datasets.}

\item{init_value}{Initial value to assign to each alpha.p parameter.}
}
\value{
A named list of initial values.
}
\description{
Creates a named list of initial values for detection intercepts per dataset.
}
\keyword{internal}
