% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building_toolkit.R
\name{build_lookups}
\alias{build_lookups}
\title{Build NIMBLE model lookups}
\usage{
build_lookups(species_datasets_list)
}
\arguments{
\item{species_datasets_list}{A named list of species datasets.}
}
\value{
A list containing \code{site_lookup}, \code{year_lookup}, and \code{species_samples}.
}
\description{
Prepares and returns a set of lookup tables and cleaned species samples for model input.
}
