% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building_toolkit.R
\name{build_model_arguments}
\alias{build_model_arguments}
\title{Build arguments for defineModel()}
\usage{
build_model_arguments(
  species_samples,
  state_covariates,
  det_covariates,
  inclPhenology = FALSE
)
}
\arguments{
\item{species_samples}{List of formatted species datasets.}

\item{state_covariates}{Named list of state covariates.}

\item{det_covariates}{Named list of detection covariates.}

\item{inclPhenology}{Logical indicating whether to include phenology terms.}
}
\value{
A list of arguments used in defineModel.
}
\description{
Prepares a list of arguments for defining the iSDM modelcode.
}
