% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building_toolkit.R
\name{build_model_constants}
\alias{build_model_constants}
\title{Build model constants}
\usage{
build_model_constants(
  species_datasets_list,
  state_covariates_list = NULL,
  ob_covariates_list = NULL,
  inclPhenology
)
}
\arguments{
\item{species_datasets_list}{A list of species datasets with site and year information.}

\item{state_covariates_list}{Optional list of state covariates.}

\item{ob_covariates_list}{Optional list of observation covariates.}

\item{inclPhenology}{Logical; should phenology constants be included?}
}
\value{
A named list of constants.
}
\description{
Constructs a named list of constants required for the nimble iSDM.
}
\details{
Includes sample dimensions, visit counts, site/year IDs, and optional phenology and covariate constants.
}
