% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building_toolkit.R
\name{build_model_inits}
\alias{build_model_inits}
\title{Build list of initial values for model}
\usage{
build_model_inits(
  species_datasets_list,
  state_covariates = NULL,
  det_covariates = NULL,
  peakDay = metadata$peakDay,
  n_datasets = length(species_datasets_list),
  inclPhenology = FALSE
)
}
\arguments{
\item{species_datasets_list}{List of species datasets.}

\item{state_covariates}{Optional list of state covariates.}

\item{det_covariates}{Optional list of detection covariates.}

\item{peakDay}{Peak day used for phenology initialisation.}

\item{n_datasets}{Number of species datasets.}

\item{inclPhenology}{Logical indicating whether to include phenology parameters.}
}
\value{
A list of named initial values.
}
\description{
Compiles initial values for latent state, core parameters, phenology, and covariate coefficients.
}
