% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building_toolkit.R
\name{build_species_samples}
\alias{build_species_samples}
\title{Build list of species samples data}
\usage{
build_species_samples(species_datasets_list, date_id = "date")
}
\arguments{
\item{species_datasets_list}{A list of species datasets with \code{x1km_grid}, \code{year}, \code{y1}, \code{siteID}, \code{yearID}, and optionally \code{date}.}

\item{date_id}{The name of the date column (default = "date").}
}
\value{
A list of cleaned species samples suitable for input to model functions.
}
\description{
Extracts relevant fields for model input from each species dataset, returning a clean list.
}
\keyword{internal}
