% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_toolkit.R
\name{drop_covs}
\alias{drop_covs}
\title{Drop specified covariate columns from a list of data frames}
\usage{
drop_covs(cov_list, cols = c())
}
\arguments{
\item{cov_list}{A list of data frames, each containing covariate columns.}

\item{cols}{A character vector of column names to drop from each data frame in the list.}
}
\value{
A list of data frames with the specified columns removed.
}
\description{
Removes unwanted columns from each element (data frame) in a list of covariate datasets.
Useful for pruning unused or redundant covariates prior to formatting or modelling.
}
\examples{
covs <- list(
  df1 = data.frame(a = 1:3, b = 4:6, c = 7:9),
  df2 = data.frame(a = 10:12, b = 13:15, c = 16:18)
)
cleaned <- drop_covs(covs, cols = c("b", "c"))

}
