% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building_toolkit.R
\name{format_det_covariates}
\alias{format_det_covariates}
\title{Format detection covariates for use in NIMBLE models}
\usage{
format_det_covariates(
  covariates_det_list,
  impute_missing = FALSE,
  scale_cov = TRUE,
  center = FALSE
)
}
\arguments{
\item{covariates_det_list}{A named list of data frames. Each data frame should contain one or more detection covariates
for a species or dataset. Names of the list will be used in diagnostic messages.}

\item{impute_missing}{Logical. If \code{TRUE}, missing values in each covariate are imputed using the mean of the non-missing values. Default is \code{FALSE}.}

\item{scale_cov}{Logical. If \code{TRUE}, all covariates are scaled using \code{scale()}. Default is \code{TRUE}.}

\item{center}{Logical. Passed to \code{scale()}. If \code{TRUE}, variables are centered before scaling. Default is \code{FALSE}.}
}
\value{
A list of lists. Each sub-list contains numeric vectors named by covariate and dataset index
(e.g., \code{LL1}, \code{mins_on_site2}, etc.), ready for use in NIMBLE model input.
}
\description{
This function processes a list of detection covariate datasets by optionally imputing missing values,
scaling the covariates, and renaming each column to a unique name for NIMBLE compatibility.
}
\examples{
# Example with two datasets, each with 2 covariates
cov_list <- list(
  species1 = data.frame(temp = c(20, 22, NA), effort = c(5, 7, 8)),
  species2 = data.frame(temp = c(18, 19, 21), effort = c(6, 7, 7))
)
formatted <- format_det_covariates(cov_list, impute_missing = TRUE)

}
