% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building_toolkit.R
\name{get_cov_names}
\alias{get_cov_names}
\title{Extract covariate names}
\usage{
get_cov_names(cov_list, format = c("vector", "list"))
}
\arguments{
\item{cov_list}{A named list of covariates.}

\item{format}{Output format, either "vector" or "list".}
}
\value{
A vector or list of cleaned covariate names.
}
\description{
Utility function to extract and clean covariate names from a list of covariate inputs.
}
\keyword{internal}
