% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_toolkit.R
\name{get_monads}
\alias{get_monads}
\title{Convert eastings/northings to OSGB 1km grid squares}
\usage{
get_monads(data, crs = "OSGB")
}
\arguments{
\item{data}{A data frame with columns \code{x} and \code{y} (eastings/northings).}

\item{crs}{Character. Currently only "OSGB" supported.}
}
\value{
A data frame with new \code{x1km_grid} column.
}
\description{
Uses \code{BRCmap} functions to convert x/y coordinates to 1km OS grid references (monads).
}
