% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_toolkit.R
\name{import_climate_rasters}
\alias{import_climate_rasters}
\title{Import climate covariates}
\usage{
import_climate_rasters(
  get_monads = FALSE,
  region_filepath = NULL,
  gr_col = "grid_ref"
)
}
\arguments{
\item{get_monads}{Logical. Whether to compute x1km_grid from coordinates.}

\item{region_filepath}{Optional CSV of grid refs to filter datasets.}

\item{gr_col}{Column in region file containing grid references.}
}
\value{
A list of climate datasets: \code{climate_nonspatial}, \code{climate_xy}, \code{soil_data}.
}
\description{
Imports processed raster-derived climate covariates. Can extract 1km grid references if \code{get_monads = TRUE}.
}
