% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_toolkit.R
\name{import_landcover_dfs}
\alias{import_landcover_dfs}
\title{Import and process landcover covariates}
\usage{
import_landcover_dfs(
  filepath = "/data/env-data/ukceh_landcover_all_aux_robB.csv",
  region_filepath = NULL,
  gr_col = "grid_ref"
)
}
\arguments{
\item{filepath}{Path to CSV file of landcover data.}

\item{region_filepath}{Optional CSV file with grid references for filtering.}

\item{gr_col}{Name of grid ref column in region file.}
}
\value{
A list with covariate data frames.
}
\description{
Imports UKCEH landcover data and performs centre + scaling transformations.
}
