% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_toolkit.R
\name{import_species_data}
\alias{import_species_data}
\title{Import pre-processed species datasets}
\usage{
import_species_data(datadir = "/data/model-data/")
}
\arguments{
\item{datadir}{A character string specifying the directory containing the \code{.rds} files.
Default is \code{"/data/model-data/"}.}
}
\value{
A named list of species datasets (data frames).
}
\description{
Reads all \code{.rds} files from a specified directory and returns them as a named list of data frames,
each representing a separate species dataset.
}
\details{
This function assumes that each file in the directory is a pre-cleaned RDS object with a consistent structure.
}
\examples{
\dontrun{
species_data <- import_species_data(datadir = "myfolder/cleaned-data/")
}

}
