% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_toolkit.R
\name{match_nearest_env}
\alias{match_nearest_env}
\title{Match missing environmental data nearest available OSGB square}
\usage{
match_nearest_env(data, values, sitecodes, yearcodes = NULL, dims = 1)
}
\arguments{
\item{data}{Environmental data.}

\item{values}{Variable to extract from data.}

\item{sitecodes}{Data frame of site names / grid references.}

\item{yearcodes}{Optional year names to match.}

\item{dims}{1 (site) or 2 (site-year).}
}
\value{
An environmental dataset matched to survey sites and/or years.
}
\description{
Identifies missing x1km grids (and optionally years) and substitutes the nearest available environmental information.
}
