% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_toolkit.R
\name{nimblefy_i}
\alias{nimblefy_i}
\title{Format environmental covariate data for NIMBLE (1D, site)}
\usage{
nimblefy_i(data, dataset_name, values, sitecodes)
}
\arguments{
\item{data}{Environmental data with \code{x1km_grid}.}

\item{dataset_name}{Name of the covariate dataset.}

\item{values}{Name of variable to extract.}

\item{sitecodes}{Data frame with \code{x1km_grid} and \code{siteID}.}
}
\value{
A data frame with columns \code{siteID} and \code{value}.
}
\description{
Matches grid references to siteID and returns a site-level covariate vector.
}
