% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_toolkit.R
\name{nimblefy_it}
\alias{nimblefy_it}
\title{Format environmental covariate data for NIMBLE (2D, site/year)}
\usage{
nimblefy_it(data, dataset_name, values, sitecodes, yearcodes)
}
\arguments{
\item{data}{Environmental data with \code{x1km_grid} and \code{year}.}

\item{dataset_name}{Name of the covariate dataset.}

\item{values}{Name of variable to extract.}

\item{sitecodes}{Data frame with \code{x1km_grid} and \code{siteID}.}

\item{yearcodes}{Data frame with \code{year} and \code{yearID}.}
}
\value{
A matrix with dimensions siteID × yearID.
}
\description{
Reshapes a site-by-year covariate matrix from long to wide using siteID and yearID.
}
