% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_toolkit.R
\name{nimblefy_t}
\alias{nimblefy_t}
\title{Format environmental covariate data for NIMBLE (1D, year)}
\usage{
nimblefy_t(data, dataset_name, values, yearcodes)
}
\arguments{
\item{data}{Environmental data with \code{year}.}

\item{dataset_name}{Name of the covariate dataset.}

\item{values}{Name of variable to extract.}

\item{yearcodes}{Data frame with \code{year} and \code{yearID}.}
}
\value{
A data frame with columns \code{yearID} and \code{value}.
}
\description{
Matches years to yearID and returns a year-level covariate vector.
}
