% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_toolkit.R
\name{prep_nimble_cov}
\alias{prep_nimble_cov}
\title{Wrapper to prepare environmental covariates for NIMBLE}
\usage{
prep_nimble_cov(
  data_list,
  dataset_name,
  values,
  sitecodes = NULL,
  yearcodes = NULL,
  dims = 1,
  matching = FALSE,
  nimblefy_method = "i"
)
}
\arguments{
\item{data_list}{Named list of environmental data frames.}

\item{dataset_name}{Name of the target dataset within \code{data_list}.}

\item{values}{Name of variable to extract.}

\item{sitecodes}{Data frame with \code{x1km_grid} and \code{siteID}.}

\item{yearcodes}{Data frame with \code{year} and \code{yearID}.}

\item{dims}{Either 1 or 2 for site/year or site-year data.}

\item{matching}{Logical. Whether to use nearest-match fallback.}

\item{nimblefy_method}{"i", "t", or "it" for site, time, or site-year structure.}
}
\value{
Covariate formatted for input to NIMBLE.
}
\description{
Handles filtering, matching, and formatting of covariate data for site, year, or site-year structures.
}
