% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_toolkit.R
\name{valid_gb_monads}
\alias{valid_gb_monads}
\title{Create lookup of valid GB monads}
\usage{
valid_gb_monads(env_filepath, env_var_name = "tas_MAT")
}
\arguments{
\item{env_filepath}{Path to the \code{.rds} file containing the environmental raster stack.}

\item{env_var_name}{Name of the variable to extract from the list (default is "tas_MAT").}
}
\value{
A data frame of 1km monads from the specified environmental variable.
}
\description{
Loads background environmental data and applies the \code{get_monads()} function to extract the spatial lookup.
}
\keyword{internal}
